Reimplement debugger window tiler (#5)
This commit reimplements the Cog visual debugger's window tiler from
the ground up. The new implementation improves on the old in a couple
important ways: it can reuse the desktop space made available by a
window closing, and its internal representation of the desktop area
enables it to find and use regions of free space that the old tiler
couldn't or didn't exploit.

The old tiler subdivided and kept track of the desktop area using
a binary tree. This frequently led to irregularly sized and aligned
fragments that were difficult to fit windows into. It also had no
support for merging freed areas back together - as the tiler was used
its representation of the desktop would become increasingly fragmented
and its effectiveness at placing new windows would decrease.

The new tiler's internal represation of the desktop is a grid of cells
maintained in a sort of 2D linked list. Cells, or rather whole columns
or rows of cells, can be split to maintain a nice, regular grid where
all cells in a column have the same width and all cells in a row have
the same height. The regular, linked structure makes it easy to inspect
a cell's neighbors to see if a component can span across them. This in
turn makes it unecessary to try and merge regions of free space back
together when a component is removed from the desktop. The end result
is a tiler that more efficiently uses the space available and
continues to do so even as its internal representation becomes
increasingly fragmented.